ocr/NuMarkdown-8B-Thinking/requirements.txt
gradio
vllm[triton]
Pillow
pytesseract
requests
aiohttp
PyMuPDF
